Transform sanity check into a test
