+++ package.json
@@ -17,8 +17,7 @@
     "pretty": "prettier -w \"**/*.ts\"",
     "playground": "pnpm --dir packages/playground build",
     "playground:clean": "pnpm --dir packages/playground clean",
-    "watch": "tsc -b tsconfig.build.json --watch",
-    "sanity-check:token-enum-mapping": "tsx check-token-to-enum-mapping.ts"
+    "watch": "tsc -b tsconfig.build.json --watch"
   },
   "devDependencies": {
     "@types/node": "^18.19.113",
